requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
docker==6.1.3
Flask==2.3.2
frozenlist==1.3.3
GDAL==3.6.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
packaging==23.1
pandas==2.0.1
Pillow==9.4.0
plotly==5.14.1
protobuf==4.21.12
psycopg2==2.9.6
pycares==4.3.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
websocket-client==1.6.0
Werkzeug==2.3.4
yarl==1.9.2
